requirements.txt
distlib==0.3.9
filelock==3.17.0
mysql-connector-python==9.3.0
platformdirs==4.3.6
python-dotenv==1.1.1
virtualenv==20.29.2
